tools/sensor_sim/requirements.txt
asyncio-mqtt
paho-mqtt<2
click
PyYAML
